Improve the `sink` handling in `getTextContent` for Highlight annotations (PR 20019 follow-up)
Currently there's a couple of issues related to the `sink` handling:
 - The `Page.prototype.extractTextContent` method is invoked with options that it doesn't actually use; note https://github.com/mozilla/pdf.js/blob/1ddf6449ac4e5c8157249dd0cd584112104fdd60/src/core/document.js#L669-L676

 - When parsing "nested" textContent, i.e. /Form /XObjects, we end up wrongly treating repeated /XObjects as empty for the annotations use-case since `enqueue` is never invoked; note https://github.com/mozilla/pdf.js/blob/1ddf6449ac4e5c8157249dd0cd584112104fdd60/src/core/evaluator.js#L3439 and https://github.com/mozilla/pdf.js/blob/1ddf6449ac4e5c8157249dd0cd584112104fdd60/src/core/evaluator.js#L3449-L3451

 - The `getTextContent` method might become ever so slightly slower by having to defer parsing at every step, given the "bad" fallback value when comparing with the `TEXT_CONTENT_CHUNK_SIZE` constant (in the API), note https://github.com/mozilla/pdf.js/blob/1ddf6449ac4e5c8157249dd0cd584112104fdd60/src/display/api.js#L1705 and https://github.com/mozilla/pdf.js/blob/1ddf6449ac4e5c8157249dd0cd584112104fdd60/src/core/evaluator.js#L3566

 - Having the `sink` now be effectively optional, in the `getTextContent` method, does complicate the code slightly overall.

To address these things this patch ensures that a `sink` will always be available, by re-using the `sinkWrapper` structure from the "nested" textContent case, and with reasonable default values.
